fix docs warnings
move docs for plug_opts to top-level
